requirements.txt
# Core
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0

# Guardrails AI (Condition A)
guardrails-ai>=0.6.0
pydantic>=2.0.0

# NeMo Guardrails (Condition B)
nemoguardrails>=0.11.0

# Data & Analysis
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.14.0
matplotlib>=3.9.0
seaborn>=0.13.0
tabulate>=0.9.0

# Async / Concurrency
aiohttp>=3.10.0
asyncio-throttle>=1.0.0
tqdm>=4.66.0

# Dataset loading
datasets>=3.0.0
jsonlines>=4.0.0
